Create file structure for project
Merging this PR will create a file structure for the Ruby project
